New migration to edit the 'users' table fields, and add check for 'user_permissions' migration to avoid migration errors

+++ src/database/migrations/2021_05_31_010233_create_user_permissions_table.php
@@ -14,16 +14,18 @@ class CreateUserPermissionsTable extends Migration
      */
     public function up()
     {
-        Schema::create('user_permissions', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade')->onUpdate('cascade');
-            $table->string('permission')->default('use-app');;
-            $table->string('restrictions')->nullable();
-            $table->timestamp('created_at')->useCurrent();
-            $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
+        if (!Schema::hasTable('user_permissions')) {        // check existence first (to ensure compatibility with pre-existing projects)
+            Schema::create('user_permissions', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade')->onUpdate('cascade');
+                $table->string('permission')->default('use-app');;
+                $table->string('restrictions')->nullable();
+                $table->timestamp('created_at')->useCurrent();
+                $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
 
-            $table->unique(['user_id', 'permission']);
-        });
+                $table->unique(['user_id', 'permission']);
+            });
+        }
     }
 
     /**
